+++ encore_api_cli/exceptions.py
@@ -1,0 +1,10 @@
+class ClientError(Exception):
+    pass
+
+
+class InvalidFileType(ClientError):
+    pass
+
+
+class RequestsError(ClientError):
+    pass